% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bincounter_cpp}
\alias{bincounter_cpp}
\title{count occurances in bins. Useful for power calculations. Replaces hist command from R.}
\usage{
bincounter_cpp(x, bins)
}
\arguments{
\item{x}{numeric vector}

\item{bins}{numeric vector}
}
\value{
Integer vector of counts
}
\description{
count occurances in bins. Useful for power calculations. Replaces hist command from R.
}
\keyword{internal}
