% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binner_cont}
\alias{binner_cont}
\title{routine to bin continuous data and assure expected counts>5}
\usage{
binner_cont(
  x,
  pnull,
  param,
  k = 10L,
  which = 1L,
  Range = as.numeric(c(-99999, 99999))
)
}
\arguments{
\item{x}{A numeric vector of numbers.}

\item{pnull}{function to find CDF}

\item{param}{parameter of pnull}

\item{k}{=10, number of bins}

\item{which}{=1, 1=equal probability bins, 2-equal size bins}

\item{Range}{=(-99999,99999) limits of possible observations, if any}
}
\value{
A vectors of bin edges
}
\description{
routine to bin continuous data and assure expected counts>5
}
\keyword{internal}
