% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binner_disc}
\alias{binner_disc}
\title{routine to bin discrete data and assure expected counts>5}
\usage{
binner_disc(x, p, k = 10L, minexpcount = 2)
}
\arguments{
\item{x}{An integer vector of counts.}

\item{p}{cumulative distribution function values at vals}

\item{k}{number of desired bins}

\item{minexpcount}{=2 minimal expected bin count required}
}
\value{
An integer vector of indices
}
\description{
routine to bin discrete data and assure expected counts>5
}
\keyword{internal}
