% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chi_stat_cont}
\alias{chi_stat_cont}
\title{chi square statistic for continuous data}
\usage{
chi_stat_cont(param, x, pnull, bins, formula, rate = 0)
}
\arguments{
\item{param}{Vector of parameters of pnull}

\item{x}{Vector of observations}

\item{pnull}{Function to calculate probabilities (pmf)}

\item{bins}{bins to use}

\item{formula}{type of chi square formula to use}

\item{rate}{rate of Poisson if sample size is random, 0 otherwise}
}
\value{
chi square statistic
}
\description{
chi square statistic for continuous data
}
\keyword{internal}
