% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chi_stat_disc}
\alias{chi_stat_disc}
\title{chi square statistic for discrete data}
\usage{
chi_stat_disc(param, x, pnull, bins, formula, rate = 0)
}
\arguments{
\item{param}{parameter for pnull}

\item{x}{Vector of observations}

\item{pnull}{cumulative distribution function under the null hypothesis}

\item{bins}{vector of indices of bins}

\item{formula}{type of chi square formula to use}

\item{rate}{rate of Poisson if sample size is random, 0 otherwise}
}
\value{
chi square statistic
}
\description{
chi square statistic for discrete data
}
\keyword{internal}
