% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chi_test_disc}
\alias{chi_test_disc}
\title{This function performs a number of chisquare gof tests for discrete data}
\usage{
chi_test_disc(
  x,
  pnull,
  param,
  nbins = as.integer(c(100, 10)),
  formula = "Pearson",
  rate = 0,
  Minimize = 0L,
  minexpcount = 2
)
}
\arguments{
\item{x}{data set}

\item{pnull}{cdf under the null hypothesis}

\item{param}{starting values of multi-D minimum chi square minimization}

\item{nbins}{=c(100, 10) number of bins for chisquare tests}

\item{formula}{type of chi square formula to use}

\item{rate}{rate of Poisson if sample size is random, 0 otherwise}

\item{Minimize}{Should minimum chi square be found?}

\item{minexpcount}{=2 minimal expected bin count required}
}
\value{
A numeric matrix of test statistics, degrees of freedom and p values
}
\description{
This function performs a number of chisquare gof tests for discrete data
}
\keyword{internal}
