% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gof_cont}
\alias{gof_cont}
\title{run gof tests for continuous data}
\usage{
gof_cont(
  x,
  pnull,
  phat,
  rnull,
  qnull,
  B = 5000L,
  doMethod = as.character(c("KS", "K", "AD", "CvM", "W", "ZA", "ZK", "ZC", "Wassp1"))
)
}
\arguments{
\item{x}{A numeric vector of data}

\item{pnull}{R function (cdf)}

\item{phat}{function to set or estimate parameters of pnull}

\item{rnull}{R function (generate data under null hypothesis)}

\item{qnull}{R function (quantiles under null hypothesis)}

\item{B}{(=5000) Number of simulation runs}

\item{doMethod}{List methods to include}
}
\value{
A matrix of numbers
}
\description{
run gof tests for continuous data
}
\keyword{internal}
