% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nm_calc}
\alias{nm_calc}
\title{calculate a matrix of numbers needed for Zhangs tests for discrete data}
\usage{
nm_calc(n)
}
\arguments{
\item{n}{An integer.}
}
\value{
A matrix of numbers
}
\description{
calculate a matrix of numbers needed for Zhangs tests for discrete data
}
\keyword{internal}
