% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_cont}
\alias{power_cont}
\title{find power of gof tests for continuous data}
\usage{
power_cont(
  pnull,
  rnull,
  qnull,
  ralt,
  param_alt,
  phat,
  TS,
  nbins = as.integer(c(100, 10)),
  rate = 0,
  Range = as.numeric(c(-99999, 99999)),
  B = as.integer(c(1000, 1000)),
  alpha = 0.05
)
}
\arguments{
\item{pnull}{R function (cdf)}

\item{rnull}{R function (generate data under null hypothesis)}

\item{qnull}{R function (quantiles under null hypothesis)}

\item{ralt}{R function to generate data under alternative}

\item{param_alt}{parameters of ralt}

\item{phat}{function to estimate parameters from the data}

\item{TS}{function to calculate test statistics}

\item{nbins}{=c(100,10) number of bins to use}

\item{rate}{rate of Poisson if sample size is random}

\item{Range}{=(-99999, 99999) limits of possible observations, if any}

\item{B}{=c(1000, 1000) Number of simulation runs for power and null distribution}

\item{alpha}{=0.05, type I error of test}
}
\value{
A matrix of powers
}
\description{
find power of gof tests for continuous data
}
\keyword{internal}
