% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_cont}
\alias{TS_cont}
\title{Find test statistics for continuous data}
\usage{
TS_cont(
  x,
  Fx,
  param,
  qnull,
  doMethod = as.character(c("KS", "K", "AD", "CvM", "W", "ZA", "ZK", "ZC", "Wassp1"))
)
}
\arguments{
\item{x}{A numeric vector.}

\item{Fx}{numeric vector of cdf probabilities.}

\item{param}{parameters for pnull}

\item{qnull}{An R function, the quantile function under the null hypothesis.}

\item{doMethod}{A character vector of methods to include}
}
\value{
A numeric vector with test statistics
}
\description{
Find test statistics for continuous data
}
\keyword{internal}
