% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_disc}
\alias{TS_disc}
\title{Find test statistics for discrete data}
\usage{
TS_disc(
  x,
  p,
  nm,
  vals,
  doMethod = as.character(c("KS", "K", "AD", "CvM", "W", "ZA", "ZK", "ZC", "Wassp1"))
)
}
\arguments{
\item{x}{An integer vector.}

\item{p}{A numeric vector of probabilities.}

\item{nm}{A matrix of pre-calculated (with nm_calc.cpp) numbers needed for Zhangs tests.}

\item{vals}{A numeric vector with the values of the discrete rv.}

\item{doMethod}{A character vector of methods to include}
}
\value{
A vector with test statistics
}
\description{
Find test statistics for discrete data
}
\keyword{internal}
