% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gof_disc}
\alias{gof_disc}
\title{run gof tests for discrete data}
\usage{
gof_disc(x, pnull, rnull, vals, phat, TS, rate = 0, B = 5000L)
}
\arguments{
\item{x}{an integer vector of counts}

\item{pnull}{cumulative distribution function under the null hypothesis}

\item{rnull}{R function (generate data under null hypothesis)}

\item{vals}{numeric vector of values of discrete random variables.}

\item{phat}{function to estimate parameters}

\item{TS}{function that calculates test statistics}

\item{rate}{=0 rate of Poisson if sample size is random, 0 otherwise.}

\item{B}{(=5000) Number of simulation runs}
}
\value{
A matrix of numbers
}
\description{
run gof tests for discrete data
}
\keyword{internal}
